Fix 'implicit conversion loses integer precision' warnings on OS X
- for assignment to member vars just cast rather than change
  type as that's less intrusive. For local vars change the
  type.

+++ Core/Contents/Include/PolyScene.h
@@ -120,7 +120,7 @@ namespace Polycode {
 		bool isEnabled();		
 		void setEnabled(bool enabled);
 		
-		int getNumEntities() { return entities.size(); }
+		int getNumEntities() { return (int)entities.size(); }
 		SceneEntity *getEntity(int index) { return entities[index]; }
 		
 		/**
